Add a method to remove the private part of a key
To enable a keypair to be encoded without exposing the private part,
provide a method to return the same key with just the public
information. This is a no-op if there is not private part to the key.

This also adds methods to the builders that copy the values from an
existing key.